form.go replaced by main.go, validation added to store KV form
