+++ wiki/ConsulWithGoBlog/index.html
@@ -6,8 +6,8 @@
     <div class="container">
         <h4>Store or Get Key/Value Pairs</h4>
         <form action="" method="post">
-            <input type="text" name="key">
-            <input type="text" name="value">
+            <input type="text" name="key" required>
+            <input type="text" name="value" required>
             <input type="submit" name="store-submit" value="Store KV pair">
         </form>
         <br>
